requirements.txt
PyMuPDF
spacy
pytesseract
pdf2image
Pillow
Flask